requirements.txt
aiofiles==0.7.0
altgraph==0.17.3
anyio==4.3.0
certifi==2024.2.2
chardet==5.2.0
colorama==0.4.6
colorlog==6.7.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
Markdown==3.6
numpy==1.26.4
pefile==2023.2.7
pillow==10.2.0
psutil==5.9.8
pydantic==1.10.14
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.7
pypiwin32==223
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
starlette==0.37.1
ttkbootstrap==1.10.1
typing_extensions==4.9.0
websockets==10.4
wxPython==4.2.1
yiri-mirai==0.2.7
zipp==3.19.1
